removes a warning.
git-svn-id: https://emgucv.svn.sourceforge.net/svnroot/emgucv/trunk@1579 d7f09016-e345-0410-b530-edf29a71df78

+++ Emgu.CV.Extern/gpu/gpuHOGDescriptor.cpp
@@ -61,5 +61,5 @@ void gpuHOGDescriptorDetectMultiScale(
 
    descriptor->detectMultiScale(*img, rects, hitThreshold, winStride, padding, scale, groupThreshold);
    if (!rects.empty())
-      cvSeqPushMulti(foundLocations, &rects[0], rects.size());
+      cvSeqPushMulti(foundLocations, &rects[0], static_cast<int>(rects.size()));
 }
